Refactor HIP backend (#5410)
* Move HIP DeepCopy to it's own files

* Remove useless macro guard

* Move HIP SharedAllocationRecord to its own files

* Move Kokkos::HIP out of Kokkos_HIP_Space and into Kokkos_HIP

* Move Kokkos_HIP.hpp and Kokkos_HIP_Space.hpp to HIP

* Fix indentation

* Add missing include

* Remove non-public header guard in Kokkos_HIP.hpp and Kokkos_HIP_Space.hpp

* Fix macro guard
