+++ core/src/HIP/Kokkos_HIP_Instance.cpp
@@ -52,8 +52,8 @@
 #include <Kokkos_Core.hpp>
 
 #include <HIP/Kokkos_HIP_Instance.hpp>
-#include <Kokkos_HIP.hpp>
-#include <Kokkos_HIP_Space.hpp>
+#include <HIP/Kokkos_HIP.hpp>
+#include <HIP/Kokkos_HIP_Space.hpp>
 #include <impl/Kokkos_Error.hpp>
 
 /*--------------------------------------------------------------------------*/
