+++ core/src/HIP/Kokkos_HIP_Space.hpp
@@ -42,17 +42,10 @@
 //@HEADER
 */
 
-#ifndef KOKKOS_IMPL_PUBLIC_INCLUDE
-#include <Kokkos_Macros.hpp>
-static_assert(false,
-              "Including non-public Kokkos header files is not allowed.");
-#endif
 #ifndef KOKKOS_HIPSPACE_HPP
 #define KOKKOS_HIPSPACE_HPP
 
 #include <Kokkos_Core_fwd.hpp>
-
-#if defined(KOKKOS_ENABLE_HIP)
 
 #include <iosfwd>
 #include <typeinfo>
@@ -61,15 +54,14 @@ static_assert(false,
 #include <iosfwd>
 
 #include <Kokkos_HostSpace.hpp>
-#include <Kokkos_Layout.hpp>
 #include <Kokkos_ScratchSpace.hpp>
 #include <HIP/Kokkos_HIP_Error.hpp>  // HIP_SAFE_CALL
 
 #include <impl/Kokkos_Profiling_Interface.hpp>
 #include <impl/Kokkos_HostSharedPtr.hpp>
-#include <impl/Kokkos_InitializationSettings.hpp>
 
 #include <hip/hip_runtime_api.h>
+
 /*--------------------------------------------------------------------------*/
 
 namespace Kokkos {
@@ -375,378 +367,7 @@ struct MemorySpaceAccess<HIPManagedSpace, HIPHostPinnedSpace> {
   enum : bool { deepcopy = true };
 };
 
-};  // namespace Impl
-//----------------------------------------
-
-}  // namespace Kokkos
-
-/*--------------------------------------------------------------------------*/
-/*--------------------------------------------------------------------------*/
-
-namespace Kokkos {
-namespace Impl {
-
-void DeepCopyHIP(void* dst, const void* src, size_t n);
-void DeepCopyAsyncHIP(const HIP& instance, void* dst, const void* src,
-                      size_t n);
-void DeepCopyAsyncHIP(void* dst, const void* src, size_t n);
-
-template <class MemSpace>
-struct DeepCopy<MemSpace, HostSpace, HIP,
-                std::enable_if_t<is_hip_type_space<MemSpace>::value>> {
-  DeepCopy(void* dst, const void* src, size_t n) { DeepCopyHIP(dst, src, n); }
-  DeepCopy(const HIP& instance, void* dst, const void* src, size_t n) {
-    DeepCopyAsyncHIP(instance, dst, src, n);
-  }
-};
-
-template <class MemSpace>
-struct DeepCopy<HostSpace, MemSpace, HIP,
-                std::enable_if_t<is_hip_type_space<MemSpace>::value>> {
-  DeepCopy(void* dst, const void* src, size_t n) { DeepCopyHIP(dst, src, n); }
-  DeepCopy(const HIP& instance, void* dst, const void* src, size_t n) {
-    DeepCopyAsyncHIP(instance, dst, src, n);
-  }
-};
-
-template <class MemSpace1, class MemSpace2>
-struct DeepCopy<MemSpace1, MemSpace2, HIP,
-                std::enable_if_t<is_hip_type_space<MemSpace1>::value &&
-                                 is_hip_type_space<MemSpace2>::value>> {
-  DeepCopy(void* dst, const void* src, size_t n) { DeepCopyHIP(dst, src, n); }
-  DeepCopy(const HIP& instance, void* dst, const void* src, size_t n) {
-    DeepCopyAsyncHIP(instance, dst, src, n);
-  }
-};
-
-template <class MemSpace1, class MemSpace2, class ExecutionSpace>
-struct DeepCopy<MemSpace1, MemSpace2, ExecutionSpace,
-                std::enable_if_t<is_hip_type_space<MemSpace1>::value &&
-                                 is_hip_type_space<MemSpace2>::value &&
-                                 !std::is_same<ExecutionSpace, HIP>::value>> {
-  inline DeepCopy(void* dst, const void* src, size_t n) {
-    DeepCopyHIP(dst, src, n);
-  }
-
-  inline DeepCopy(const ExecutionSpace& exec, void* dst, const void* src,
-                  size_t n) {
-    exec.fence(fence_string());
-    DeepCopyAsyncHIP(dst, src, n);
-  }
-
- private:
-  static const std::string& fence_string() {
-    static const std::string string =
-        std::string("Kokkos::Impl::DeepCopy<") + MemSpace1::name() + "Space, " +
-        MemSpace2::name() +
-        "Space, ExecutionSpace>::DeepCopy: fence before copy";
-    return string;
-  }
-};
-
-template <class MemSpace, class ExecutionSpace>
-struct DeepCopy<MemSpace, HostSpace, ExecutionSpace,
-                std::enable_if_t<is_hip_type_space<MemSpace>::value &&
-                                 !std::is_same<ExecutionSpace, HIP>::value>> {
-  inline DeepCopy(void* dst, const void* src, size_t n) {
-    DeepCopyHIP(dst, src, n);
-  }
-
-  inline DeepCopy(const ExecutionSpace& exec, void* dst, const void* src,
-                  size_t n) {
-    exec.fence(fence_string());
-    DeepCopyAsyncHIP(dst, src, n);
-  }
-
- private:
-  static const std::string& fence_string() {
-    static const std::string string =
-        std::string("Kokkos::Impl::DeepCopy<") + MemSpace::name() +
-        "Space, HostSpace, ExecutionSpace>::DeepCopy: fence before copy";
-    return string;
-  }
-};
-
-template <class MemSpace, class ExecutionSpace>
-struct DeepCopy<HostSpace, MemSpace, ExecutionSpace,
-                std::enable_if_t<is_hip_type_space<MemSpace>::value &&
-                                 !std::is_same<ExecutionSpace, HIP>::value>> {
-  inline DeepCopy(void* dst, const void* src, size_t n) {
-    DeepCopyHIP(dst, src, n);
-  }
-
-  inline DeepCopy(const ExecutionSpace& exec, void* dst, const void* src,
-                  size_t n) {
-    exec.fence(fence_string());
-    DeepCopyAsyncHIP(dst, src, n);
-  }
-
- private:
-  static const std::string& fence_string() {
-    static const std::string string =
-        std::string("Kokkos::Impl::DeepCopy<HostSpace, ") + MemSpace::name() +
-        "Space, ExecutionSpace>::DeepCopy: fence before copy";
-    return string;
-  }
-};
 }  // namespace Impl
 }  // namespace Kokkos
 
-//----------------------------------------------------------------------------
-//----------------------------------------------------------------------------
-
-namespace Kokkos {
-namespace Impl {
-
-template <>
-class SharedAllocationRecord<HIPSpace, void>
-    : public HostInaccessibleSharedAllocationRecordCommon<HIPSpace> {
- private:
-  friend class SharedAllocationRecordCommon<HIPSpace>;
-  friend class HostInaccessibleSharedAllocationRecordCommon<HIPSpace>;
-  using base_t     = HostInaccessibleSharedAllocationRecordCommon<HIPSpace>;
-  using RecordBase = SharedAllocationRecord<void, void>;
-
-  SharedAllocationRecord(const SharedAllocationRecord&) = delete;
-  SharedAllocationRecord& operator=(const SharedAllocationRecord&) = delete;
-
-#ifdef KOKKOS_ENABLE_DEBUG
-  static RecordBase s_root_record;
-#endif
-
-  const HIPSpace m_space;
-
- protected:
-  ~SharedAllocationRecord();
-
-  template <typename ExecutionSpace>
-  SharedAllocationRecord(
-      const ExecutionSpace& /*exec*/, const HIPSpace& arg_space,
-      const std::string& arg_label, const size_t arg_alloc_size,
-      const RecordBase::function_type arg_dealloc = &base_t::deallocate)
-      : SharedAllocationRecord(arg_space, arg_label, arg_alloc_size,
-                               arg_dealloc) {}
-
-  SharedAllocationRecord(
-      const HIP& exec_space, const HIPSpace& arg_space,
-      const std::string& arg_label, const size_t arg_alloc_size,
-      const RecordBase::function_type arg_dealloc = &base_t::deallocate);
-
-  SharedAllocationRecord(
-      const HIPSpace& arg_space, const std::string& arg_label,
-      const size_t arg_alloc_size,
-      const RecordBase::function_type arg_dealloc = &base_t::deallocate);
-};
-
-template <>
-class SharedAllocationRecord<HIPHostPinnedSpace, void>
-    : public SharedAllocationRecordCommon<HIPHostPinnedSpace> {
- private:
-  friend class SharedAllocationRecordCommon<HIPHostPinnedSpace>;
-  using base_t     = SharedAllocationRecordCommon<HIPHostPinnedSpace>;
-  using RecordBase = SharedAllocationRecord<void, void>;
-
-  SharedAllocationRecord(const SharedAllocationRecord&) = delete;
-  SharedAllocationRecord& operator=(const SharedAllocationRecord&) = delete;
-
-#ifdef KOKKOS_ENABLE_DEBUG
-  static RecordBase s_root_record;
-#endif
-
-  const HIPHostPinnedSpace m_space;
-
- protected:
-  ~SharedAllocationRecord();
-  SharedAllocationRecord() = default;
-
-  template <typename ExecutionSpace>
-  SharedAllocationRecord(
-      const ExecutionSpace& /*exec_space*/, const HIPHostPinnedSpace& arg_space,
-      const std::string& arg_label, const size_t arg_alloc_size,
-      const RecordBase::function_type arg_dealloc = &base_t::deallocate)
-      : SharedAllocationRecord(arg_space, arg_label, arg_alloc_size,
-                               arg_dealloc) {}
-
-  SharedAllocationRecord(
-      const HIPHostPinnedSpace& arg_space, const std::string& arg_label,
-      const size_t arg_alloc_size,
-      const RecordBase::function_type arg_dealloc = &base_t::deallocate);
-};
-
-template <>
-class SharedAllocationRecord<HIPManagedSpace, void>
-    : public SharedAllocationRecordCommon<HIPManagedSpace> {
- private:
-  friend class SharedAllocationRecordCommon<HIPManagedSpace>;
-  using base_t     = SharedAllocationRecordCommon<HIPManagedSpace>;
-  using RecordBase = SharedAllocationRecord<void, void>;
-
-  SharedAllocationRecord(const SharedAllocationRecord&) = delete;
-  SharedAllocationRecord& operator=(const SharedAllocationRecord&) = delete;
-
-#ifdef KOKKOS_ENABLE_DEBUG
-  static RecordBase s_root_record;
-#endif
-
-  const HIPManagedSpace m_space;
-
- protected:
-  ~SharedAllocationRecord();
-  SharedAllocationRecord() = default;
-
-  template <typename ExecutionSpace>
-  SharedAllocationRecord(
-      const ExecutionSpace& /*exec_space*/, const HIPManagedSpace& arg_space,
-      const std::string& arg_label, const size_t arg_alloc_size,
-      const RecordBase::function_type arg_dealloc = &base_t::deallocate)
-      : SharedAllocationRecord(arg_space, arg_label, arg_alloc_size,
-                               arg_dealloc) {}
-
-  SharedAllocationRecord(
-      const HIPManagedSpace& arg_space, const std::string& arg_label,
-      const size_t arg_alloc_size,
-      const RecordBase::function_type arg_dealloc = &base_t::deallocate);
-};
-}  // namespace Impl
-}  // namespace Kokkos
-
-//----------------------------------------------------------------------------
-//----------------------------------------------------------------------------
-
-namespace Kokkos {
-namespace Impl {
-class HIPInternal;
-}
-/// \class HIP
-/// \brief Kokkos device for multicore processors in the host memory space.
-class HIP {
- public:
-  //------------------------------------
-  //! \name Type declarations that all Kokkos devices must provide.
-  //@{
-
-  //! Tag this class as a kokkos execution space
-  using execution_space = HIP;
-  using memory_space    = HIPSpace;
-  using device_type     = Kokkos::Device<execution_space, memory_space>;
-
-  using array_layout = LayoutLeft;
-  using size_type    = HIPSpace::size_type;
-
-  using scratch_memory_space = ScratchMemorySpace<HIP>;
-
-  HIP();
-  HIP(hipStream_t stream, bool manage_stream = false);
-
-  //@}
-  //------------------------------------
-  //! \name Functions that all Kokkos devices must implement.
-  //@{
-
-  KOKKOS_INLINE_FUNCTION static int in_parallel() {
-#if defined(__HIP_DEVICE_COMPILE__)
-    return true;
-#else
-    return false;
-#endif
-  }
-
-  /** \brief Wait until all dispatched functors complete.
-   *
-   * The parallel_for or parallel_reduce dispatch of a functor may return
-   * asynchronously, before the functor completes. This method does not return
-   * until all dispatched functors on this device have completed.
-   */
-  static void impl_static_fence(const std::string& name);
-
-  void fence(const std::string& name =
-                 "Kokkos::HIP::fence(): Unnamed Instance Fence") const;
-
-  hipStream_t hip_stream() const;
-
-  /// \brief Print configuration information to the given output stream.
-  void print_configuration(std::ostream& os, bool verbose = false) const;
-
-  /// \brief Free any resources being consumed by the device.
-  static void impl_finalize();
-
-  /** \brief  Initialize the device.
-   *
-   */
-  int hip_device() const;
-  static hipDeviceProp_t const& hip_device_prop();
-
-  static void impl_initialize(InitializationSettings const&);
-
-  static int impl_is_initialized();
-
-  //  static size_type device_arch();
-
-  static size_type detect_device_count();
-
-  static int concurrency();
-  static const char* name();
-
-  inline Impl::HIPInternal* impl_internal_space_instance() const {
-    return m_space_instance.get();
-  }
-
-  uint32_t impl_instance_id() const noexcept;
-
- private:
-  friend bool operator==(HIP const& lhs, HIP const& rhs) {
-    return lhs.impl_internal_space_instance() ==
-           rhs.impl_internal_space_instance();
-  }
-  friend bool operator!=(HIP const& lhs, HIP const& rhs) {
-    return !(lhs == rhs);
-  }
-  Kokkos::Impl::HostSharedPtr<Impl::HIPInternal> m_space_instance;
-};
-
-namespace Tools {
-namespace Experimental {
-template <>
-struct DeviceTypeTraits<HIP> {
-  static constexpr DeviceType id = DeviceType::HIP;
-  static int device_id(const HIP& exec) { return exec.hip_device(); }
-};
-}  // namespace Experimental
-}  // namespace Tools
-
-namespace Impl {
-template <class DT, class... DP>
-struct ZeroMemset<HIP, DT, DP...> {
-  ZeroMemset(const HIP& exec_space, const View<DT, DP...>& dst,
-             typename View<DT, DP...>::const_value_type&) {
-    KOKKOS_IMPL_HIP_SAFE_CALL(hipMemsetAsync(
-        dst.data(), 0,
-        dst.size() * sizeof(typename View<DT, DP...>::value_type),
-        exec_space.hip_stream()));
-  }
-
-  ZeroMemset(const View<DT, DP...>& dst,
-             typename View<DT, DP...>::const_value_type&) {
-    KOKKOS_IMPL_HIP_SAFE_CALL(
-        hipMemset(dst.data(), 0,
-                  dst.size() * sizeof(typename View<DT, DP...>::value_type)));
-  }
-};
-}  // namespace Impl
-}  // namespace Kokkos
-
-namespace Kokkos {
-namespace Impl {
-
-template <>
-struct MemorySpaceAccess<HIPSpace, HIP::scratch_memory_space> {
-  enum : bool { assignable = false };
-  enum : bool { accessible = true };
-  enum : bool { deepcopy = false };
-};
-
-}  // namespace Impl
-}  // namespace Kokkos
-
-#endif /* #if defined( KOKKOS_ENABLE_HIP ) */
 #endif /* #define KOKKOS_HIPSPACE_HPP */
